Menubar: Fix crash in WM_COMMAND handling.
For not yet fully understood reason, pressing <ALT> to activate a menubar
and then (after releasing the <ALT>) pressing <SPACE>, an endless
recursion of WM_COMMAND is caused bu some interaction of a dialog or
top-level window and processing in original toolbar procedure.

We fix by not propagating WM_COMMAND if it is not generated by the menu.